move `VirtualStateApproxId` struct to `block` module
